+++ CHANGELOG.md
@@ -12,4 +12,4 @@ The **third number** is for emergencies when we need to start branches for older
 
 ### Added
 
--   Initial release of Cookiecutter Python CLI App Demo
+- Initial release of Cookiecutter Python CLI App Demo